requirements.txt
python-telegram-bot==13.15
apscheduler==3.10.4
sqlalchemy==2.0.23
pymysql==1.1.0
pytz==2023.3
openpyxl==3.1.2
